requirements.txt
fastapi[standard]==0.115.13
sqlalchemy==2.0.30
mysql-connector-python==8.4.0
python-dotenv==1.0.1
pydantic-settings==2.3.4
alembic==1.13.1
uvicorn==0.30.1
black==24.4.2
ruff==0.5.3
pytest==8.2.2
pytest-mock==3.14.0
pytest-cov==5.0.0
pytest-asyncio==1.1.0
pytest-env==1.1.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.5.0
google-generativeai==0.8.5
sendgrid==6.11.0